[REFACTOR] refact return error code
