+++ so_long/get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: junyoo <junyoo@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/01 11:42:47 by junyoo            #+#    #+#             */
-/*   Updated: 2022/12/11 18:56:14 by junyoo           ###   ########.fr       */
+/*   Updated: 2022/12/12 17:45:46 by junyoo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
