+++ so_long/map_utils.c
@@ -6,7 +6,7 @@
 /*   By: junyoo <junyoo@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/06 18:10:34 by junyoo            #+#    #+#             */
-/*   Updated: 2022/12/11 22:02:29 by junyoo           ###   ########.fr       */
+/*   Updated: 2022/12/13 13:24:44 by junyoo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
