+++ so_long/maps/big.ber
@@ -1,0 +1,20 @@
+111111111111111111111111111111111111111
+1E0000000000000C00CCCCC000C000000000001
+10100101001000001010010000CCCCC00010101
+101001001010101000100100CCCCC0000010101
+101001001010101000100100CCCCC0000010101
+101001001010101000100100CCCCC0000010101
+101001001010101000100100CCCCC0000010101
+101001001010101000100100CCCCC0000010101
+1P0000000C00C000000000000000000000000C1
+10100101001000001010010000CCCCC00010101
+100000000000000C00CCCCC000C000000000001
+10100101001000001010010000CCCCC00010101
+100000000000000C00CCCCC000C000000000001
+10100101001000001010010000CCCCC00010101
+10100101001000001010010000CCCCC00010101
+100000000000000C00CCCCC000C000000000001
+10100101001000001010010000CCCCC00010101
+100000000000000C00CCCCC000C000000000001
+10100101001000001010010000CCCCC00010101
+111111111111111111111111111111111111111
