+++ so_long/maps/blocked.ber
@@ -1,0 +1,5 @@
+1111111111111
+1C010000000C1
+1100011111001
+1P1011E0000C1
+1111111111111
