+++ so_long/so_long.c
@@ -6,7 +6,7 @@
 /*   By: junyoo <junyoo@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/06 18:00:43 by junyoo            #+#    #+#             */
-/*   Updated: 2022/12/11 22:28:34 by junyoo           ###   ########.fr       */
+/*   Updated: 2022/12/13 14:36:30 by junyoo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,12 +26,20 @@ int	main(int argc, char *argv[])
 		img_set(g);
 		map_set(g, 2, g->c_rat);
 		mlx_key_hook(g->win, deal_key, g);
-		mlx_hook(g->win, 17, 0, end_game, g);
+		mlx_hook(g->win, 17, 0, exit_game, g);
 		mlx_loop(g->mlx);
 	}
 	else
 		perror ("please input only one map");
 	return (0);
+}
+
+int	exit_game(t_game *game)
+{
+	write(1, "user exited game\n", 17);
+	mlx_destroy_window(game->mlx, game->win);
+	free(game);
+	exit(0);
 }
 
 int	end_game(t_game *game)
@@ -56,33 +64,6 @@ void	print_cat_move(int c_move, int flag)
 		write(1, "\ngame over!! your cat moved ", 28);
 	ft_putnbr(c_move);
 	write(1, " steps\n", 7);
-}
-
-void	ft_putnbr(int n)
-{
-	long long	temp;
-	int			i;
-	char		str[15];
-
-	temp = (long long)n;
-	i = 0;
-	if (n < 0)
-	{
-		temp *= -1;
-		write (1, "-", 1);
-	}
-	else if (n == 0)
-	{
-		write (1, "0", 1);
-		return ;
-	}
-	while (temp)
-	{
-		str[i++] = (temp % 10) + '0';
-		temp /= 10;
-	}
-	while (--i >= 0)
-		write(1, str + i, 1);
 }
 
 //arch -x86_64 gcc -L./mlx -lmlx -framework OpenGL -framework AppKit main.c
