--- so_long/source/get_next_line.h
@@ -1,33 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   get_next_line.h                                    :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: junyoo <junyoo@student.42seoul.kr>         +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/09/01 10:48:12 by junyoo            #+#    #+#             */
-/*   Updated: 2022/12/02 15:09:04 by junyoo           ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef GET_NEXT_LINE_H
-# define GET_NEXT_LINE_H
-# define BUFFER_SIZE	13
-
-# include <unistd.h>
-# include <stdlib.h>
-
-typedef struct s_list
-{
-	int				fd;
-	char			*text;
-	struct s_list	*next;
-}	t_list;
-
-char			*get_next_line(int fd);
-size_t			ft_strlen(const char *s);
-char			*ft_substr(char const *s, unsigned int start, size_t len);
-char			*ft_strchr(const char *s, int c);
-char			*ft_strjoin(char const *s1, char const *s2);
-
-#endif
